discord/discord_tools/requirements.txt
aeventkit==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
discord==2.3.2
discord.py==2.6.4
fastapi==0.121.3
frozenlist==1.8.0
h11==0.16.0
ib_async==2.0.1
idna==3.11
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.5
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
